% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_quarantinesim}
\alias{run_quarantinesim}
\title{Run the quarantinesim Shiny application}
\usage{
run_quarantinesim()
}
\value{
None. This function runs the Shiny app.
}
\description{
This function launches the quarantinesim Shiny app, which lets users
explore outbreak probabilities by R0, vaccine effectiveness (VE),
and vaccination coverage.
}
\examples{
if (interactive()) {
  run_quarantinesim()
}

}
